[OEMDATA] Add OEM Data build rule
Adds a new build rule for OEM data files which are structured
compiled data. The rule provides support for MSFT, GCC, and
Clang toolchains.

Deletes the intermediate binary files in a module output directory
after the final binary is written using the files. This is
important to prevent intermediate files from potentially being
used in future incremental builds if the corresponding source
file is no longer present.

BaseTools/OemDataFormatter: Rename to ProductDataFormatter

Renames the tool to ProductDataFormatter to align with a change in
data name from "OEM data" to "product data".