MdeModulePkg: Update Memory Logging
This patch skips looping through the GCD if DEBUG_GCD is not set,
as this is a performance hit to touch every GCD entry just to not
print it.

It also lowers the print level of SetUefiImageMemoryAttributes to
verbose as it is quite a noisy log. While there, it also ensures
that gCpu is available in that function before it is dereferenced.